Add old museumplus-pas Ansible configuration
Add Ansible configuration that was used for the previous deployment.

This will be used as the base for the Ansible configuration of the new
Passari deployment.

The files are copied as-is from another Git repository, but the vault
files are excluded.  They are written by Janne Pulkkinen for
Museovirasto on 2019-2020.  (Copyright holder is Museovirasto.)