Add requirements files
Add requirements.in and requirements.txt for listing/pinning the
packages needed to run Ansible.

It's also possible to use the Ansible packaged with various Linux
distros, but having a pinned version defined in requirements.txt helps
to find a version which is certainly compatible with the playbook files
here.

Also add requirements-dev.* for pinning ansible-dev-tools and its
dependencies.  Especially useful for development is ansible-lint, which
will be installed via ansible-dev-tools.