Improve Passari Web db initialization
If there was any issue with the playbook between the Passari Web venv
creation and running the handlers, then the database was never
initialized.  Make sure it is always initialized by moving the
initialization from handlers to tasks.  Use the "already initialized"
text to determine when it actually changed something.

Additionally make sure that the SElinux context is applied for the venv
by flushing the handlers right after venv creation.